1 - Now the athlete's data is sent to Team Service. 2 - I migrated all code to typescript
