+++ .env.development
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_YOUR_IP=192.168.15.4
