--- components/appPage/form/Option/index.js
@@ -1,11 +1,0 @@
-export default function Option({ text, value, hidden }) {
-	return (
-		<>
-			<option value={value} hidden={hidden}>
-				{text}
-			</option>
-			<style jsx>{`
-			`}</style>
-		</>
-	);
-}
