+++ package.json
@@ -18,10 +18,8 @@
   },
   "homepage": "https://github.com/ThallyssonKlein/Smartial-Website#readme",
   "dependencies": {
-    "@react-firebase/firestore": "^0.5.5",
     "apisauce": "^2.1.1",
     "date-fns": "^2.21.3",
-    "firebase": "^8.5.0",
     "next": "^10.0.3",
     "public-ip": "^4.0.3",
     "pubnub": "^4.32.0",
@@ -29,5 +27,11 @@
     "qrcode.react": "^1.0.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1"
+  },
+  "devDependencies": {
+    "@types/pubnub": "^4.29.3",
+    "@types/qrcode.react": "^1.0.2",
+    "@types/react": "^17.0.14",
+    "typescript": "^4.3.5"
   }
 }
